[Feat] Tile UI 작업 중
