+++ Assets/Modules/UI/Board.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 629b7476914deff4baf71a016ff2e4ae
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
