+++ Assets/Modules/UI/Board/Magic Level Up.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 176bae5d849dffa49b3c4170ab8e664a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
