+++ Assets/Modules/UI/Board/UITile.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Cardinals.Board {
+    public class UITile: MonoBehaviour {
+        
+    }
+}
