+++ Assets/Modules/UI/Board/UITile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1980a36c1ea84bf41ab24d3b8f4d7f80
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
