--- Assets/Modules/UI/Stage/Magic Level Up.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 176bae5d849dffa49b3c4170ab8e664a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
