--- Assets/Modules/UI/Stage/Magic Level Up/UIInitialMagicSelectPanel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 55d618e55e15ae84fbc663642187f483
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
