--- Assets/Modules/UI/Stage/Magic Level Up/UILevelUpMagicSlot.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3ec41eae6671eea4fa3e57110b235aa0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
