--- Assets/Modules/UI/Stage/Magic Level Up/UIMagicLevelUpPanel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9368b0ead9f94a2498e59ea8cb5eb456
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
