--- Assets/Modules/UI/Stage/Magic Level Up/UIPostMagicSelectPanel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a8c60df5d28aa7640922999d321fb4a0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
