+++ Assets/Resources/Sprites/UI/Icons/TileActions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 31e63b100ee66ff4f8370d1e8a505786
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
